Expand UI doc and mkdocs with certs and backends
Enhanced UI documentation to cover the management of certificates and backends. Updated mkdocs navigation to reflect the addition of a new 'Certs' section, providing a more comprehensive resource for users. This update aims to improve user guidance and the overall usability of the Prism UI documentation within the CoCoS system.

Signed-off-by: SammyOina <sammyoina@gmail.com>