Add backend management documentation and resource images
Introduced comprehensive documentation for the backends service in the project, detailing the management process of backends including creation, updating, listing, viewing, terminating, and deletion along with mutual TLS security. The documentation is accompanied by new diagrams and UI images to visualize backend configurations and interactions. These additions aim to provide clear instructions and resources for users, facilitating a better understanding of the backend lifecycle and its management within the system.

Signed-off-by: SammyOina <sammyoina@gmail.com>